backend/app.py
from flask import Flask
from flask_cors import CORS
from dotenv import load_dotenv
from routes.scan import scan_bp

load_dotenv()

app = Flask(__name__)
CORS(app)

app.register_blueprint(scan_bp)

if __name__ == "__main__":
    app.run(debug=True, port=5000)
